Fix some bugs

+++ robotino_node/include/MotorArrayROS.h
@@ -32,7 +32,7 @@ private:
 	ros::Time stamp_;
 
 	void velocitiesChangedEvent( const float* velocities, unsigned int size );
-	void positionsChangedEvent( const float* positions, unsigned int size );
+	void positionsChangedEvent( const int* positions, unsigned int size );
 	void currentsChangedEvent( const float* currents, unsigned int size );
 };
 #endif /* MOTORARRAYROS_H_ */
